# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Seatmap SaaS API
  version: "1.0.0"
  summary: Multi-tenant seating inventory, holds, allocations, ticketing and check-in.
  description: |
    The API owns venue seating inventory. It does **not** own carts, payments, tax or refunds —
    those stay in the tenant's own storefront (see `docs/adr/0001-saas-woocommerce-boundary.md`).

    ## Authentication
    Three separate schemes, deliberately not interchangeable:

    | Audience | Scheme | Notes |
    | --- | --- | --- |
    | Tenant panel / management API | `bearerAuth` (personal access token) | Scoped to the caller's tenant and role. |
    | Server-to-server (WordPress plugin) | `apiKeyAuth` + `hmacSignature` | All four `X-Seatmap-*` headers required. |
    | Embed widget (browser) | none | Public event id only; never carries a secret. |
    | Check-in device | `bearerAuth` (device token) | Scoped to named events. |

    ## HMAC signing
    `signature = hex(hmac_sha256(secret, METHOD + "\n" + PATH + "\n" + timestamp + "\n" + nonce + "\n" + hex(sha256(raw_body))))`

    `secret` is the value returned once when the key was created. The path is part of the signed
    string, so a captured signature cannot be redirected to another endpoint.

    Timestamps must be within ±300 s of server time; nonces may not repeat inside that window.
    Both conditions are enforced, so an accurate clock on the calling server is a requirement,
    not a nicety.

    ## Idempotency
...

    ## Errors
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /sites:
    get:
      operationId: listSites
      tags: [Sites]
      summary: The organiser's hosted event sites
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Site" }
    post:
      operationId: createSite
      tags: [Sites]
      summary: Create a site
      description: |
        Creating a site provisions it: a home page, an event page, a "visiting us" page, a primary
        and a footer menu, and — if `SEATMAP_SITES_DOMAIN` is configured — a free subdomain, already
        verified. A new site is `draft` until a verified domain exists.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 120 }
                theme_key: { type: string, maxLength: 40 }
                locale: { type: string, maxLength: 12 }
                timezone: { type: string }
                currency: { type: string, minLength: 3, maxLength: 3 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SiteDetail" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /sites/{site}:
    parameters:
      - $ref: "#/components/parameters/SiteId"
    get:
      operationId: getSite
      tags: [Sites]
      summary: One site with its pages, menus and domains
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SiteDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateSite
      tags: [Sites]
      summary: Rename, restyle, or publish a site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 120 }
                theme_key: { type: string, maxLength: 40 }
                locale: { type: string, maxLength: 12 }
                timezone: { type: string }
                currency: { type: string, minLength: 3, maxLength: 3 }
                status: { type: string, enum: [draft, live] }
                brand: { $ref: "#/components/schemas/SiteBrand" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SiteDetail" }
        "409":
          description: "`no_verified_domain` — a site cannot go live with no address to visit."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: "`unknown_theme`, or an invalid field."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /sites/{site}/domains/{domain}/verify:
    parameters:
      - $ref: "#/components/parameters/SiteId"
      - $ref: "#/components/parameters/DomainId"
    post:
      operationId: verifySiteDomain
      tags: [Sites]
      summary: Check the TXT record now
      description: Runs the DNS check immediately instead of waiting for the queued one. `last_error` says why a failure failed.
      responses:
        "200":
          description: Checked — read `verified`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SiteDomain" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Seatmap-Key
      description: Public key id of an API key. Never the secret.
    hmacSignature:
      type: apiKey
      in: header
      name: X-Seatmap-Signature
      description: |
        Requires `X-Seatmap-Timestamp` (unix seconds, ±300 s) and `X-Seatmap-Nonce`
        (unique within the window) alongside.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, maxLength: 255 }
...
    SiteId:
      name: site
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    DomainId:
      name: domain
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    NotFound:
      description: Not found, or not visible to this tenant
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ValidationError:
      description: Invalid payload
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable identifier.
              examples: [seat_unavailable, capacity_unavailable, hold_expired, invalid_signature, tenant_limit_reached]
            message: { type: string }
            details:
              type: object
              additionalProperties: true
            request_id: { type: string }
...
    SiteBrand:
      description: |
        The organiser's own look, inside bounds. Colours, fonts, radii and logo URLs are validated
        on the way in — a brand is a set of choices from closed lists, not arbitrary CSS, because
        this renders on a domain we serve.
      type: object
      properties:
        accent: { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
        heading_font: { type: string }
        body_font: { type: string }
        radius: { type: string }
        logo_url: { type: string, format: uri }
        tagline: { type: string, maxLength: 160 }
        gateways:
          type: array
          description: Payment drivers the storefront offers. Defaults to `offline` — pay at the box office.
          items: { type: string }
    Site:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        theme_key: { type: string }
        locale: { type: string }
        timezone: { type: string }
        currency: { type: string }
        brand: { $ref: "#/components/schemas/SiteBrand" }
        status: { type: string, enum: [draft, live] }
        url:
          type: [string, "null"]
          description: Canonical address, or null while no domain is verified.
        domains:
          type: array
          items: { $ref: "#/components/schemas/SiteDomain" }
    SiteDetail:
      allOf:
        - $ref: "#/components/schemas/Site"
        - type: object
          properties:
            pages:
              type: array
              items: { $ref: "#/components/schemas/SitePage" }
            menus:
              type: array
              items: { $ref: "#/components/schemas/SiteMenu" }
    SitePage:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        title: { type: string }
        kind:
          type: string
          enum: [home, page, event]
          description: "`home` and `event` are structural: the site needs them, so they cannot be deleted and home cannot move off the root."
        path: { type: string }
        blocks:
          type: array
          description: The draft. What visitors see is the published copy, which this response does not carry.
          items: { $ref: "#/components/schemas/Block" }
        seo_title: { type: [string, "null"] }
        seo_description: { type: [string, "null"] }
        position: { type: integer }
        published_at: { type: [string, "null"], format: date-time }
        has_unpublished_changes: { type: boolean }
    Block:
      type: object
      description: |
        One piece of a page. `type` decides which other properties are meaningful; everything else
        is dropped on the way in, so a stored block only ever holds fields its type defines.
      required: [type]
      properties:
        id:
          type: string
          description: Stable across edits, so reordering one block does not break an anchor to another.
        type:
          type: string
          enum: [heading, richText, image, buttons, eventList, eventDetail, faq, venueMap, divider, html]
      additionalProperties: true
    SiteMenu:
      type: object
      properties:
        key: { type: string, examples: [primary, footer] }
        name: { type: string }
        items:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              label: { type: string }
              target_type: { type: string, enum: [page, event, url] }
              site_page_id: { type: [string, "null"], format: uuid }
              event_id: { type: [string, "null"], format: uuid }
              url: { type: [string, "null"] }
              new_tab: { type: boolean }
    SiteDomain:
      type: object
      properties:
        id: { type: string, format: uuid }
        hostname: { type: string }
        is_primary: { type: boolean }
        verified:
          type: boolean
          description: Until this is true the hostname is claimed but not served.
        verified_at: { type: [string, "null"], format: date-time }
        last_checked_at: { type: [string, "null"], format: date-time }
        last_error: { type: [string, "null"] }
        record:
          type: object
          description: The DNS record to publish before verification can pass.
          properties:
            type: { type: string, examples: [TXT] }
            name: { type: string }
            value: { type: string }
...